+++ src/GSvar/IgvDialog.h
@@ -2,6 +2,7 @@
 #define IGVDIALOG_H
 
 #include <QDialog>
+#include <QTreeWidgetItem>
 
 namespace Ui {
 class IgvDialog;
@@ -15,14 +16,14 @@ public:
 	explicit IgvDialog(QWidget *parent = 0);
 	~IgvDialog();
 
-	void addFile(QString label, QString filename, bool checked);
-	void addSeparator();
+	void addFile(QString label, QString type, QString filename, bool checked);
 
 	bool skipForSession() const;
 	QStringList filesToLoad();
 
 protected slots:
 	void on_skip_session_clicked();
+	void treeItemChanged(QTreeWidgetItem* item);
 
 private:
 	Ui::IgvDialog *ui;