+++ src/GSvar/MainWindow.h
@@ -14,6 +14,13 @@
 #include "FileWatcher.h"
 #include "BusyDialog.h"
 
+struct IgvFile
+{
+	QString id; //sample identifier/name (for visualization)
+	QString type; //file type (for grouping)
+	QString filename; //file name
+};
+
 ///Main window class
 class MainWindow
 		: public QMainWindow
@@ -32,11 +39,11 @@ public:
 	///Returns the LOG files corresponding to the variant list.
 	QStringList getLogFiles();
 	///Returns the BAM files for the analysis.
-	QMap<QString, QString> getBamFiles();
+	QList<IgvFile> getBamFiles();
 	///Returns CNV SEG files for the analysis.
-	QMap<QString, QString> getSegFilesCnv();
+	QList<IgvFile> getSegFilesCnv();
 	///Returns BAF SEG files for the analysis.
-	QMap<QString, QString> getIgvFilesBaf();
+	QList<IgvFile> getIgvFilesBaf();
 	enum VariantListType {GERMLINE_SINGLESAMPLE, GERMLINE_TRIO, GERMLINE_MULTISAMPLE, SOMATIC_SINGLESAMPLE, SOMATIC_PAIR};
 	///Retruns the type of the current variant list
 	VariantListType getType();