+++ src/cppNGS-TEST/NGSHelper_Test.h
@@ -126,37 +126,4 @@ private slots:
 		NGSHelper::softClipAlignment(al,33038800,33038809);
 		S_EQUAL(al.cigarDataAsString(), "10S131M10S5H");
 	}
-
-	void getSampleHeader_singlesample_noheader()
-	{
-		QString input = TESTDATA("data_in/VariantFilter_in.GSvar");
-		VariantList vl;
-		vl.load(input);
-		SampleHeaderInfo info = NGSHelper::getSampleHeader(vl, input);
-		I_EQUAL(info.count(), 1);
-		I_EQUAL(info.sampleColumns().count(), 1);
-		S_EQUAL(info.sampleColumns()[0], "genotype");
-		I_EQUAL(info.sampleColumns(true).count(), 1);
-		S_EQUAL(info.sampleColumns(true)[0], "genotype");
-		I_EQUAL(info.sampleColumns(false).count(), 0);
-	}
-
-	void getSampleHeader_multisample()
-	{
-		QString input = TESTDATA("data_in/VariantFilter_in_multi.GSvar");
-		VariantList vl;
-		vl.load(input);
-		SampleHeaderInfo info = NGSHelper::getSampleHeader(vl, input);
-		I_EQUAL(info.count(), 3);
-		I_EQUAL(info.sampleColumns().count(), 3);
-		S_EQUAL(info.sampleColumns()[0], "Affected1");
-		S_EQUAL(info.sampleColumns()[1], "Affected2");
-		S_EQUAL(info.sampleColumns()[2], "Control");
-		I_EQUAL(info.sampleColumns(true).count(), 2);
-		S_EQUAL(info.sampleColumns(true)[0], "Affected1");
-		S_EQUAL(info.sampleColumns(true)[1], "Affected2");
-		I_EQUAL(info.sampleColumns(false).count(), 1);
-		S_EQUAL(info.sampleColumns(false)[0], "Control");
-	}
-
 };