+++ src/cppNGS/NGSHelper.h
@@ -4,27 +4,6 @@
 #include "cppNGS_global.h"
 #include "BamReader.h"
 
-///Sample header struct for samples in variant lists (GSvar).
-struct CPPNGSSHARED_EXPORT SampleInfo
-{
-	QString column_name;
-	QMap<QString, QString> properties;
-
-	///Returns if the sample has state 'affected'.
-	bool isAffected() const;
-};
-
-///Sample header information from GSvar files.
-class CPPNGSSHARED_EXPORT SampleHeaderInfo
-	: public QMap<QString, SampleInfo>
-{
-	public:
-		///Returns all sample genotype column names of all samples.
-		 QStringList sampleColumns() const;
-		///Returns all sample genotype column names of affected/unaffected samples.
-		QStringList sampleColumns(bool affected) const;
-};
-
 ///Helper class for NGS-specific stuff.
 class CPPNGSSHARED_EXPORT NGSHelper
 {
@@ -42,9 +21,6 @@ public:
 
 	///Create sample overview file
 	static void createSampleOverview(QStringList in, QString out, int indel_window=100, bool cols_auto=true, QStringList cols = QStringList());
-
-	///Parses and returns sample data from variant list header. The @p filename argument is needed to support old germline variant lists without sample header.
-	static SampleHeaderInfo getSampleHeader(const VariantList& vl, QString gsvar_file);
 
 	///Expands a Amino acid notation with 1 letter to 3 letters
 	static QByteArray expandAminoAcidAbbreviation(QChar amino_acid_change_in);